Ensure composer installs it in the correct path
Composer on the bedrock docker build keeps putting it in `/app/vendor`,
which it should be going in to `/app/web/app/themes`.  I *think* this is
because its `composer.json` had it set as `"type":"library"`. This is an
experimental push to see if changing the type fixes the problem.