Allow placement in /tmp on docker
Setting this to `wordpress-theme` prevents me from overriding the
bedrock default `install-path` on docker. I'm changing it back to
`library` for this reason and because, techncially, it isn't a theme. It
*contains* a theme, but that won't work unless it is extracted from this
repo and put in the correct location--something I cannot do if the
entire repo is being clone into `./themes`.