Ensure it works with new and existing containers
Cleaning the config files from the root path was causing breakages on
subsequent runs of the same container.

I have also removed the composer auth code for now for brevity  I will
put it back once we have the CI environment running if it is needed.
Examples can be found it most TP WP repos.