Use a smaller, more current mailcatcher container
The one we were using is two minor Ruby versions behind and weighs
in at around 260Mb.  Moving this to a reference version of the Ruby alpine
container and only installing the minimum dependencies brings it
up-to-date and down in size to around 90Mb.

I have left in the nginx proxy settings (the `VIRTUAL_...` envs), in the
docker compose file although they don't appear to be strictly necessary
if the ports are exposed and it is only run locally.  However, this may
not be the case if we are going to use mailcatcher to test in remote CI
environments.  # Please enter the commit message for your changes. Lines
starting