Update the WP CLI script to not remove users
Instead of removing and recreating an already existing user, we are just
going to update their password. This way we can maintain whatever setup the
user had configured (like agencies, regions, etc.) as unfortunately wp-cli
doesn't seem to support these custom fields.

Also, if an ENV variable `pass_admin` exists, it will also create/update an
Administrator user.

As wp-cli can't seem to support setting the agencies and regions of the test
users (specifically user `agency_editor`) we need to login with an admin user
and manually asign said test user to an agency for many tests to pass.