Redirect all requests for /wp-admin correctly
For reasons that are not clear, `location /wp-admin {...` is refusing to
match paths with filenames.  This means that login paths like
`/wp-admin/post-new.php?post_type=post` do not redirect correctly with
the original configuration.  Adding an explicit prefix-matching
operator (`location ^~ /wp-admin {...` and contextualising the rewrite
rule to the block (`rewrite_rule ^/wp-admin...` -> `rewrite_rule
^/(.*)...`) fixes this problem.