Fix gateway errors that result from FPM timeouts
The 502/504 errors we are seeing on the AWS prod enviroment seem to stem
from the very conservative fpm settings from the base container.  After
discussing with DS, we agreed to raise the limits.  I am starting with
5 processes and 15 processes maximum.  I am also setting a respawn limit
of 1500 requests in case DXW were dealing with memory leaks that we were
not made aware of.