Make logs searchable; Enable fpm slow log
It appears cloudwatch has very limited search capabilities for non-json
formatted logs–simple wildcards only. JSON formatted logs, however,
have a much more sophsiticated search DSL.  This should make it much
easier to find things in the logs.

I have also enabled the fpm slow log in an effort to track down the
cause of the remaining 504s and slow event feed load times.  The setting
I've chosen are aggressive–they may need review after a few days.  Also,
it is worth noting that triggering this log locally is tricky.  As a
result, I do not know its format, so it may be diffcult to search on
cloudwatch. There does not seem to be a simple way to configure it to
output json.  I will experiment with doing so if it becomes necessary.