Ensure caching meets expectations of end users
The Content team expect that newly published articles will appear on the
site within 60 seconds of publication.  Due to the aggressive nature of
browser caching, and our varying `cache_timeout` settings, this was
frequently not the case.  This change ensures that all caching will be
done for 60 seconds.

I started by making all of the `cache_timeout` settings equal to 60,
then I realised this could be DRY'd out by simply removing them and
setting the default timeout to 60 seconds.  I replaced the
`cache_timeout` key for error pages, only-set to zero seconds, this
continues to prevent the error pages from caching by sending max-age=0;

This *may* cause performance problem on production, so we should monitor
server performance for some time after it is released.