Prepare smoke tests to integrate with Travis
Prepare the smoke tests container to be able to run on CI against the
wordpress container, and setup users, etc.
Cleanup of git-cryp, env files, and other stuff not really needed anymore.

WIP: right now it is running against an RDS database instance, and only
running a dummy test. The next step would be to get rid of the RDS instance
and to have a local mariaDB container loading a dump of the DB, as the RDS
solution turned out to be veeeery slow. Once done, instead of the profile
`quick` we can use the default profile to run a bigger amount of tests.

https://www.pivotaltracker.com/story/show/151883867