More tweaks, exit codes and quick run
Several things here:

* In order to speed up the feedback loop and fail faster until we tune
everything, only 1 sily test will run, not the whole suite.

* The script that waits until wordpress is up and running modified to
just check for a web response, even if it is a 500.

* Saving the exit code from the `make smoketest` to return it after the
shutting down of the containers, so Travis knows if it is a success or
a failure (although this doesn't seem to work as well as it sounded in
my head).