Enable stack shutdown from test container
This wasn't working as desired using the old method.  I reduced the size
and complexity of the smoke test container and installed the
dependencies needed to talk to the docker daemon.

This commit also now launches/uses a mariadb container.  As of this
commit, this container WILL NOT WORK on travis; it still requires a db
dump to be pulled externally and made available at build time. The next
commit will cover this.