Remove image size restrictions
This is causing agency editors a lot of problems. These problems are
having a knock-on effect on the team in that we are getting requests to
upload/manage banners, which the editors should be able to do for
themselves.  Removing these restrictions should solve that problem at
the potential risk of introducing issues with mis-sized images.
However, that is an editorial issue.