Removed copy re. 3 hour limit on commenting links
The links have been observed to be active for more than three hours.
Given that access is whitelisted, and the commentor is required to have
a `justice.gov.uk` address, the security risk of the timeout **not**
functioning is very low.  As per our pre-Christmas planning, I am
removing the copy concerning the expiry from the app.