Use current version of bedrock and fix dependency
`bedrock.json` is a copy of the bedrock github repo's `composer.json`. I
chose not to use the release version as it only bumps WP to 4.8.2,
whereas this version bumps to 4.9.1. Composer is failing a dependency
check around `composer/installers`, so I'm removing that altogether to
see if it fixes it in this package.  If it does, I'll make the same
change to the other repos that are also throwing errors.

I am going to make the branch change in `moj.json` in a separate commit
as I want to be able to fiddle with that in isolation