Remove the agency switcher feature
This isn't a failing test per se; it relies on an admin building the
page and associating with a template.  It *should* have a test, but will
need some additional supporting code to ensure everything is always
built before the tests run.  Without an the associated page, the test
will fail because the app will return 404.  It will not be immediately
obvious why this happened.