Update relevanssi-premium plugin version
The old version is breaking the standard API.  We can't turn it off
because it enables us to search inside various document formats, and we
can't use the free version as it does not support this feature, either.
I've updated the premium verion in our private composer repo and updated
the composer configs here so it pull the most current version.