Refactor to solve phantomjs/docker dependencies
Phantom won't build properly on alpine and docker is difficult to
install on Debian.  However, between the two, the docker problem is e
straightforward to solve, so I changed the smoketest container to
Debian.

I did not build a custom container as I wanted to stay with official
ruby containers in order to avoid additional maintenance overhead.