Fail the entire run if supporting containers fail
This will prevent the race condition seen on travis whereby one of the
supporting containers fails to start (in that case it was the mariadb
container), but the smoketest container carries on in a wait-state.  On
travis I believe the timeout for something like this is 30 minutes.

By checking that the supporting containers are all alive at the same
time we check for WP readiness, we ensure that the tests will fail
within ten seconds of any other container failling.