Memoize the database call using the options as a key
According to New Relic, the `wp posts` table gets called 200+ times for
every request for `index.php`. It *appears* that the majority of these
calls are duplicates. This is an attempt to cut down on them and
decrease the overall processing time of `index.php`.  However, it may
not have the desired effect.