Use clarity branch with the memoized get_raw call
The idea here is that we can test this locally first by changing branch,
then merge it to master to test on dev remotely.  Finally, we can
release to prod to do a canary test there.  If that works, then the
clarity branch can be merged to master and branch in `moj.json` changed
back to master.  We can also simply change it to master if local
tests are conclusive.