Reduce max_children and max_requests
I **supect** that the 502 errors we are seeing are the result of a
mismatch between the number of nginx workers and the fpm workers.  I'm
putting the number of fpm workers back down to 20, and reducing their
lifespan (2000 requests to 1500 requests before a worker is replaced)
to see if that stops these errors.