Enable docker compose, tidy files, use supervisord
I'm trying to make the intention of all of this clearer while working
within the frame of an official php container. Ideally, I'd like to
eventually be able to move to sepearate containers for php and nginx.
For now, though, the fat-container approach is easier to port the old
configuration into.