Update nginx and consolidate configuration files
This is getting close to being releaseable. After looking over the files
and attempting a spike to separate the containers out as per docker best
practice, I decided that the configuration files could be consolidated
and simplified.  There is no need for complex `/sites-available` and
`/php-fpm.d/*` configurations when the container is never going to
serve more than a single site.

Splitting the cotainers out via fpm and nginx roles failed owing to the
tight javascript coupling between the php and the javascript and due to
issues I encountered where the php was raising errors on a standalone
fpm container. This is not happening on the consolidated fpm container.
I suspect that might be because I experimented using an alpine linux fpm
container (not gcc/glib), but I didn't have time to bottom the issue for
certain.

A sensible next step in devops terms would be to take the time needed to
split the containers. That is outside of the scope of this excersize,
however. It will be easier to do once all the repos are consolidated
into the single intranet repo as that will allow us to divide the code
by language and concern, rather than by wordpress theme.

TODO:
  * Port over the cron jobs.  This probably requires the wp-cli.
  * Ensure yas3fs is started and working.
  * Have Ollie PR the container.