Remove the request_terminate_timeout setting
I believe this is what is causing the current spate of 502 errors.
The documentation
(http://php.net/manual/en/install.fpm.configuration.php) strongly
suggests it overrides `max_execution_time`, which is set to 30 seconds.
If it does, then that would explain why we are seeing so many 502 errors
since I changed the fpm configurations.  It would also explain why it is
nearly impossible to upload a document right now-a typical document is
very, very likely to take more than 5 seconds to upload.