Move dw-markdown into main project
There was some history for this, but I have not imported it.  This is
becasue it consisted almost entirely of one-line comments describing the
change(s) made in the broadest possible terms and were not useful
The previous container was using the `update-composer-json` branch, but
I rebased that into master before pulling it in.  That branch is
probably no longer relevant as this plugin will no longer be installed
by composer.