Install config files from mappable /bedrock dir
The source directories used to both mount volumes, and to souce
configuration and code for COPY or ADD operations, should mirror the
acutal working structure of a normal bedrock-based application.

I built this container in a hurry, so this is the start of a refactor
that ensures this is the case.  It should also make the Dockerfile
easier to reuse.