Extract dw_agency cookie value and add to cache path
When caching is activated and the user changes agency, most of the pages
get stuck with the wrong agency header.  I suspect the problem lies in
the fastcgi cache on nginx because two different users making sequential
requests will both recieve the same, incorrect, header.  This will
always be the header of the first user's request.  As the fastcgi cache
*seems* to be obeying `Cache-Control` headers, and there is no other
intermediate cache (it only caches on the end-users' browsers or nginx
on the local containers), I surmise the wrong information must be
getting stored in the fastcgi cache.  This *should* fix the problem.