Send all php logs to stdout
Right now, we only get the nginx logs on the AWS enviroments. This is
causing problems when trying to track down certain issues. These
settings *should* ensure that the PHP logs are sent to CloudWatch as
well. The `access.log` and `error_log` settings were taken from the
official php container and the `slow_log` setting was interpreted.