Homepage redesign - two feature item layout (#176)
* external services added

* reduce spacing in beta banner

* black banner bar reduced

* Create popular content section on homepage

This section sits in the bottom of the homepage. It has a popular link
and blog feature.

* Add compiled CSS

* Updated CSS version

* Add compiled css

* Fix spacing on popular content section & add title

* Format PHP files

* Modified header title

* Remove video, banner and image carousel from home

* Sync up ACF fields

Adding in the new external field category plus removing many of the old
homepage items, video, polls etc to cause ACF to be out of sync. This
commit is now with everything updated and working.

* Apply version two homepage redesign

Header has been changed in ACF. ACF has undergone major changes. Blog row and news row added. Event
calendar added to popular content area. Work still needs to be done to
feature section. Third feature needs to be added.

* Implement three feature item homepage design

New homepage has been created. This includes a feature section, a latest
news section, latest blog section and more on the intranet section.
Popular in HR has been added and side banner moved to bottom. Events has
been slightly redesigned to feature a similar look to team events. A new
section in the backend has been created for the new external links that
sit within the header.

* Fix single news page aside news list

When redesigning the homepage it had a knock on effect for the layout of
the news listed causing it to look jumbled. It is now fixed and lists
out correctly. To avoid future confusion I have created a new component.

* Update CSS

* Homepage redesign - two feature item layout

After discussion with stakeholders, the three column layout has been
changed to a two column feature layout. More on the intranet section has
also been changed around a bit. I have also made improvements to the
backend of WP ACF, improving naming and layout for editors.

* Update CSS