Change rewrite rules affecting old team pages
Team pages had an unusual structure, starting with blog/. This made no
sense. It also had further legacy tree odities with a random blank
directory added in called /digital-technology/. This has all been
restructured so that it starts from root of the site with no blog and
without the digital-technology section. The changes here are redirects
to help redirect people to the new section.