Remove opt-in/out & agency bug default to hq
Slash and burn some of the opt-in stuff but more to remove. I wasn't too
particular about it as it will be refactored in upcoming commits. Also
fixed agency bug where if (as an agency) I selected another agency that
was not HQ and saved the post, it would lock both agencies out and
default the post to HQ.