Fix and update TotalPoll plugin support
TotalPolls completely refactored their poll plugin, breaking some of the
customisation work we had applied to the plugin. For example hiding
paid-for features of the plugin we are not using, in the editor view. A
warning was also thrown as the plugin had changed file names. This has
been fixed.