Change regional news template so it's available for admins
For people with Admin privileges they couldn't set the regional news
archive template from the backend template dropdown. This was a problem
because not only could admins not set the correct template in the
regional section but it was causing the page that needs that template
(regional news update page) to not display correctly.