Refactors ACF fields with new JAC additon
What was on prod and what was in Git seemed to be mismatched. I made a
copy of the PHP ACF and JSON ACF from prod, loaded it to my local
machine and readded all the JAC (plus OSPT since it was missing) fields.
I think everything is up to date now.