QA tagging amend (#881)
* QA tagging amend

Should fix error:

```
The workflow is not valid. .github/workflows/integration.yml (Line: 20, Col: 17): Invalid input, pipeline is not defined in the referenced workflow.
```

* Update build.yml (#882)

* Update build.yml

In context of this issue: https://github.com/ministryofjustice/intranet/actions/runs/16940274570/workflow

It seems the workflow is missing the input declaration from the error output. What do you think of this?

* Update build.yml

* Update build.yml

* Amend duplicate tag

-t $REGISTRY/$REPOSITORY:$PIPELINE_TAG

* Bump

* Bump - testing all-tags

* Update build and deploy to include pipeling in image tag

---------

Co-authored-by: Damien Wilson <damien.wilson@digital.justice.gov.uk>